examples/auto-research-agent/requirements.txt
endee
langchain-groq
langchain-core
tavily-python
python-dotenv
spacy